+++ src/commands/mod.rs
@@ -5,4 +5,5 @@ pub mod init;
 pub mod list;
 pub mod pull;
 pub mod push;
+pub mod remove;
 pub mod shared;